refactor(fb15b): extract prepareSystemMessages + chunkText
Into bedrock-converse-utils. 38 bedrock tests pass; bedrock 980->~955.
